+++ src/richie/apps/courses/admin.py
@@ -214,7 +214,7 @@ class CourseAdmin(FrontendEditableAdminMixin, PageExtensionAdmin):
     """Admin class for the Course model"""
 
     list_display = ["title", "is_listed"]
-    frontend_editable_fields = ("code", "duration", "effort", "is_listed")
+    frontend_editable_fields = ("code", "duration", "pace", "is_listed")
 
     # pylint: disable=no-self-use
     def title(self, obj):