+++ src/richie/apps/courses/defaults.py
@@ -325,10 +325,10 @@ PAGES_INFO.update(getattr(settings, "PAGES_INFO", {}))
 
 ROOT_REVERSE_IDS = PAGES_INFO.keys()
 
-# Fields effort and duration
+# Fields pace and duration
 MINUTE, HOUR, DAY, WEEK, MONTH = "minute", "hour", "day", "week", "month"
 DEFAULT_TIME_UNIT = HOUR
-DEFAULT_EFFORT_UNIT = WEEK
+DEFAULT_PACE_UNIT = WEEK
 DEFAULT_REFERENCE_UNIT = MONTH
 
 TIME_UNITS = {