+++ tests/apps/core/test_fields_effort.py
@@ -1,31 +1,31 @@
-"""Test suite for Richie's EffortField Django models field."""
+"""Test suite for Richie's PaceField Django models field."""
 from unittest import mock
 
 from django.db import models
 from django.test import TestCase
 
-from richie.apps.core.fields.effort import EffortField
+from richie.apps.core.fields.pace import PaceField
 
 
 @mock.patch.object(models.CharField, "check", return_value=[])
-class CheckEffortFieldTestCase(TestCase):
-    """Test suite for the `EffortField` class."""
+class CheckPaceFieldTestCase(TestCase):
+    """Test suite for the `PaceField` class."""
 
-    def test_fields_effort_time_units_required(self, _mock_check):
+    def test_fields_pace_time_units_required(self, _mock_check):
         """The `time_units` attribute is required on the field."""
-        field = EffortField()
+        field = PaceField()
         errors = field.check()
         self.assertEqual(len(errors), 1)
         error = errors[0]
         self.assertEqual(
-            error.msg, "Effort fields must define a 'time_units' attribute."
+            error.msg, "pace fields must define a 'time_units' attribute."
         )
         self.assertEqual(error.obj, field)
         self.assertEqual(error.id, "fields.E1010")
 
-    def test_fields_effort_time_units_tuple(self, _mock_check):
-        """The `time_units` object passed to instantiate an effort field should be a dictionary."""
-        field = EffortField(time_units=(("minute", "minute"),))
+    def test_fields_pace_time_units_tuple(self, _mock_check):
+        """The `time_units` object passed to instantiate an pace field should be a dictionary."""
+        field = PaceField(time_units=(("minute", "minute"),))
 
         errors = field.check()
         self.assertEqual(len(errors), 1)
@@ -35,12 +35,12 @@ class CheckEffortFieldTestCase(TestCase):
         self.assertEqual(error.obj, field)
         self.assertEqual(error.id, "fields.E1011")
 
-    def test_fields_effort_time_units_dictionary_string(self, _mock_check):
+    def test_fields_pace_time_units_dictionary_string(self, _mock_check):
         """
         The `time_units` object should list singular/plural string tuples for each possible
         time unit choice.
         """
-        field = EffortField(time_units={"minute": "minute"})
+        field = PaceField(time_units={"minute": "minute"})
 
         errors = field.check()
         self.assertEqual(len(errors), 1)
@@ -50,20 +50,20 @@ class CheckEffortFieldTestCase(TestCase):
         self.assertEqual(error.obj, field)
         self.assertEqual(error.id, "fields.E1011")
 
-    def test_fields_effort_time_units_dictionary_success(self, _mock_check):
-        """Successfuly instantiate an effort field with a valid definition of time_units."""
-        field = EffortField(time_units={"minute": ("minute", "minutes")})
+    def test_fields_pace_time_units_dictionary_success(self, _mock_check):
+        """Successfuly instantiate an pace field with a valid definition of time_units."""
+        field = PaceField(time_units={"minute": ("minute", "minutes")})
 
         errors = field.check()
         self.assertEqual(len(errors), 0)
 
-    def test_fields_effort_default_effort_unit_invalid(self, _mock_check):
+    def test_fields_pace_default_pace_unit_invalid(self, _mock_check):
         """
-        Trying to instantiate an effort field with an invalid default effort unit should
+        Trying to instantiate an pace field with an invalid default pace unit should
         not pass the checks.
         """
-        field = EffortField(
-            time_units={"minute": ("minute", "minutes")}, default_effort_unit="invalid"
+        field = PaceField(
+            time_units={"minute": ("minute", "minutes")}, default_pace_unit="invalid"
         )
 
         errors = field.check()
@@ -74,21 +74,21 @@ class CheckEffortFieldTestCase(TestCase):
         self.assertEqual(error.obj, field)
         self.assertEqual(error.id, "fields.E1012")
 
-    def test_fields_effort_default_effort_unit_success(self, _mock_check):
-        """Successfully instantiating an effort field with a default effort unit."""
-        field = EffortField(
-            time_units={"minute": ("minute", "minutes")}, default_effort_unit="minute"
+    def test_fields_pace_default_pace_unit_success(self, _mock_check):
+        """Successfully instantiating an pace field with a default pace unit."""
+        field = PaceField(
+            time_units={"minute": ("minute", "minutes")}, default_pace_unit="minute"
         )
 
         errors = field.check()
         self.assertEqual(len(errors), 0)
 
-    def test_fields_effort_default_reference_unit_invalid(self, _mock_check):
+    def test_fields_pace_default_reference_unit_invalid(self, _mock_check):
         """
-        Trying to instantiate an effort field with an invalid default reference unit should
+        Trying to instantiate an pace field with an invalid default reference unit should
         not pass the checks.
         """
-        field = EffortField(
+        field = PaceField(
             time_units={"minute": ("minute", "minutes")},
             default_reference_unit="invalid",
         )
@@ -101,9 +101,9 @@ class CheckEffortFieldTestCase(TestCase):
         self.assertEqual(error.obj, field)
         self.assertEqual(error.id, "fields.E1013")
 
-    def test_fields_effort_default_reference_unit_success(self, _mock_check):
-        """Successfully instantiating an effort field with a default reference unit."""
-        field = EffortField(
+    def test_fields_pace_default_reference_unit_success(self, _mock_check):
+        """Successfully instantiating an pace field with a default reference unit."""
+        field = PaceField(
             time_units={"minute": ("minute", "minutes")},
             default_reference_unit="minute",
         )